DAN-33 FE: Fixed an issue where device data was not being correctly fetched and converted in the provider

+++ heart_beat_client/lib/models/mood_tracking.dart
@@ -2,7 +2,7 @@ class MoodTracking {
   final int id;
   final String source;
   final String mood;
-  final int score;
+  final double score;
   final DateTime timestamp;
 
   MoodTracking({
@@ -18,7 +18,7 @@ class MoodTracking {
       id: json['id'] as int,
       source: json['source'] as String,
       mood: json['mood'] as String,
-      score: json['score'] as int,
+      score: json['score'] as double,
       timestamp: DateTime.parse(json['timestamp']),
     );
   }
